requirements.txt
python-telegram-bot>=22.0
sqlalchemy>=2.0.36
openai>=1.50.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
python-dotenv==1.0.1
reportlab==4.1.0
matplotlib>=3.8.0
openpyxl==3.1.2
fuzzywuzzy==0.18.0
python-Levenshtein>=0.23.0
apscheduler==3.10.4
pyyaml==6.0.1
tenacity==8.2.3
httpx>=0.27.0
python-dateutil>=2.8.2
pytz>=2024.1
pytest>=7.0.0